Generate data source dropin from credentials
This is a stop-gap when the sources are not defined by the quality
report. We should make a better form-based solution for this later, but
this should at least allow us to discover repositories with the data we
do have. It only functions for GitLab when there is no quality report
name.