Make configuration and export data available
Only shown if the data is placed in separate directories as indicated by
the environment variables. Links are provided on the index (to show
latest scrae) and after saving the configuration (for inspection).

Files in the directories are always shown as plain files in directory to
save people from downloading the files.