Jenkins: Adjust agent labels
Require 'axe' label to perform accessibility tests on a proper node.

This is mostly to consolidate such tests to another node, where Axe
tests for visualizations were already being done. While this test does
not have its own requirements, the dependency check done afterward is
not properly supported on all CPU architectures (zlib incompatibility),
which is the primary reason for placing them on the same node.